fix: proper teacher query
